Добавил точку входа и подключил модули
