+++ js/data.js
@@ -1,0 +1,26 @@
+const TITLES = ['Привет', 'Hello', 'Bonjour', 'die Empfehlung'];
+const APARTMENTS = ['palace', 'flat', 'house', 'bungalow'];
+const CHECKIN = ['12:00', '13:00', '14:00'];
+const CHECKOUT = ['12:00', '13:00', '14:00'];
+const DESCRIPTIONS = [
+  'Надо брать',
+  'Хорошее предложение',
+  'Подумайте хорошенько',
+  'Кошмар',
+];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+const OFFER_COUNT = 10;
+
+export {TITLES, APARTMENTS, CHECKIN, CHECKOUT, DESCRIPTIONS, FEATURES, PHOTOS, OFFER_COUNT};
